+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-03-15
+
+### Added
+
+* Surface energy script
+* Property extraction script
+
+### Removed
+
+* Point defect energy script
 
 ## [0.1.2] - 2020-10-04